requirements.txt
paho-mqtt
chromadb
requests
sentence-transformers
langchain
langchain-community
langchain-core
langchainhub